Move Home, About, and Catalog code into separate files
